Ulepszenie wylogowania

+++ ComputerShop/Client/Helpers/UserHelper.cs
@@ -1,31 +1,55 @@
 ﻿using Blazored.LocalStorage;
 using Blazored.Toast.Services;
+using ComputerShop.Client.Services.User;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
 
 namespace ComputerShop.Client.Helpers
 {
-    public static class UserHelper
+    public interface IUserHelper
     {
-        static public async Task LogOutOnUnauthorized(ILocalStorageService? localStorageService, 
-            AuthenticationStateProvider? stateProvider, IToastService? toastService, NavigationManager? navigationManager)
+        Task LogoutAsync();
+        Task LogoutOnUnauthorizedAsync();
+        Task LogoutAsync(string message = "", string title = "", string path = "", bool isError = false);
+    }
+    public class UserHelper : IUserHelper
+    {
+        private readonly AuthenticationStateProvider? stateProvider;
+        private readonly NavigationManager? navigationManager;
+        private readonly ILocalStorageService? localStorageService;        
+        private readonly IToastService? toastService;        
+        private readonly IUserService? userService;
+
+        public UserHelper(AuthenticationStateProvider? stateProvider, IUserService? userService,
+            ILocalStorageService? localStorageService, NavigationManager? navigationManager, IToastService? toastService )
         {
-            await LogOut(localStorageService, stateProvider);
-                toastService?.ShowError("Nie możemy potwierdzić twojej tożsamości", "Zaloguj się ponownie");
+            this.stateProvider = stateProvider;
+            this.navigationManager = navigationManager;
+            this.localStorageService = localStorageService;
+            this.toastService = toastService;
+            this.userService = userService;
+        }
+
+        public async Task LogoutOnUnauthorizedAsync()
+        {
+            await LogoutAsync();
+            toastService?.ShowError("Nie możemy potwierdzić twojej tożsamości", "Zaloguj się ponownie");
             navigationManager?.GoToLoginPage();
         }
-        static public async Task LogOut(ILocalStorageService? localStorageService, AuthenticationStateProvider? stateProvider, 
-            IToastService? toastService, NavigationManager? navigationManager, string message = "", string title = "", string path = "", bool isError = false)
+        public async Task LogoutAsync(string message = "", string title = "", string path = "", bool isError = false)
         {
-            await LogOut(localStorageService, stateProvider);
+            await LogoutAsync();
             if(!isError)
                 toastService?.ShowInfo(message, title);
             else
                 toastService?.ShowError(message, title);
             navigationManager?.NavigateTo(path);
         }
-        static public async Task LogOut(ILocalStorageService? localStorageService, AuthenticationStateProvider? stateProvider)
+        public async Task LogoutAsync()
         {
+            if(userService == null)
+                throw new ArgumentNullException(nameof(userService));
+            await userService.Logout();
             if (localStorageService == null)
                 throw new ArgumentNullException(nameof(localStorageService));
             await localStorageService.RemoveItemAsync("jwt");
